+++ weblate/trans/tests/test_reports.py
@@ -37,6 +37,8 @@ COUNTS_DATA = [{
     'words': 2,
     'words_edit': 0,
     'words_new': 2,
+    'chars_edit': 0,
+    'chars_new': 14,
     'email': 'weblate@example.org'
 }]
 
@@ -171,14 +173,14 @@ class ReportsTest(ViewTestCase):
             '<table>\n'
             '<tr><th>Name</th><th>Email</th>'
             '<th>Words total</th><th>Count total</th>'
-            '<th>Words edited</th><th>Count edited</th>'
-            '<th>Words new</th><th>Count new</th>'
+            '<th>Chars new</th><th>Words new</th><th>Count new</th>'
+            '<th>Chars edited</th><th>Words edited</th><th>Count edited</th>'
             '</tr>'
             '\n'
             '<tr>\n<td>Weblate Test</td>\n'
             '<td>weblate@example.org</td>\n'
             '<td>2</td>\n<td>1</td>\n'
-            '<td>2</td>\n<td>1</td>\n'
-            '<td>0</td>\n<td>0</td>\n'
+            '<td>14</td>\n<td>2</td>\n<td>1</td>\n'
+            '<td>0</td>\n<td>0</td>\n<td>0</td>\n'
             '\n</tr>\n</table>'
         )