+++ docs/changes.rst
@@ -9,6 +9,7 @@ Released on ? 2019.
 * Added support for XLIFF placeholders.
 * Celery can now utilize multiple task queues.
 * Added suppport for renaming and moving projects and components.
+* Include chars counts in reports.
 
 weblate 3.3
 -----------